+++ packages/react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @baunilha/react
 
+## 2.0.2
+
+### Patch Changes
+
+- Add fixed build and initial native support
+- Updated dependencies
+  - @baunilha/tokens@1.0.7
+
+## 2.0.1
+
+### Patch Changes
+
+- Add initial native components support
+- Updated dependencies
+  - @baunilha/tokens@1.0.6
+
 ## 2.0.0
 
 ### Major Changes