+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @baunilha/tokens
 
+## 1.0.7
+
+### Patch Changes
+
+- Add fixed build and initial native support
+
+## 1.0.6
+
+### Patch Changes
+
+- Add initial native components support
+
 ## 1.0.5
 
 ### Patch Changes