+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @baunilha/utils
 
+## 2.0.2
+
+### Patch Changes
+
+- Add fixed build and initial native support
+
+## 2.0.1
+
+### Patch Changes
+
+- Add initial native components support
+
 ## 2.0.0
 
 ### Major Changes